backend/pyproject.toml
[project]
name = "cortex"
version = "0.1.0"
description = "FastAPI Backend with Vector Search and Graph DB"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "qdrant-client>=1.12.0",
    "neo4j>=5.25.0",
    "groq>=0.11.0",
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "httpx>=0.27.0",
    "sentence-transformers>=3.2.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.18",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "ollama>=0.6.1",
    "psycopg2-binary>=2.9.7",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.1.0",
]

[tool.setuptools] 
packages = ["app"]
